Move describeGallery to its own module
